requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Flask-Login==0.6.2
python-dotenv==1.0.1
PyJWT==2.9.0
WTForms==3.0.1
email-validator==2.1.1
